+++ AsteroidGame/VisualObjects/Bullet.cs
@@ -11,6 +11,8 @@ namespace AsteroidGame.VisualObjects
     {
         private const int __BulletSizeX = 20;
         private const int __BulletSizeY = 5;
+        
+
         public Bullet(int Position)
             : base(new Point(0, Position), Point.Empty, new Size(__BulletSizeX, __BulletSizeY))
         {
